scraper-python/requirements.txt
# requirements.txt
selenium==4.24.0
undetected-chromedriver==3.1.5.post4
scrapy==2.5.0
pyyaml==6.0.2
structlog==21.1.0
tenacity==8.0.1